- Added animation preview in Prefab Mode with controls in the AnimatorAuthoring inspector. - Implemented wireframe visualization for the animated model in the Scene View to aid debugging. - Improvement: Enhanced editor resource management for GraphicsBuffer to prevent memory leaks.  -Improvement: Refactored AnimatorAuthoringEditor for better readability, structure, and performance. - Fix: Ensured the animation frame in the Scene View updates correctly when changed from the inspector.

+++ Runtime/Scripts/Components/BlobAnimator.cs
@@ -11,7 +11,7 @@ namespace SnivelerCode.GpuAnimation.Runtime.Components
 
     public struct BlobAnimatorAsset
     {
-        public int MatricesHash;
+        public ulong MatricesHash;
         public byte BoneCount;
         public BlobArray<BlobAnimationAsset> Animations;
         public uint TriggerMask;
@@ -45,6 +45,6 @@ namespace SnivelerCode.GpuAnimation.Runtime.Components
     {
         public BlobArray<float3x4> MatricesLbs;
         public BlobArray<uint> Offsets;
-        public BlobArray<int> Hashes;
+        public BlobArray<ulong> Hashes;
     }
 }
